--- src/main/java/gift/global/exception/InvalidRequestException.java
@@ -1,15 +1,0 @@
-package gift.global.exception;
-
-public class InvalidRequestException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public InvalidRequestException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
-    }
-}
